Phase 4b: extract deriveProfessorViewModel from ProfessorPanel
Move the ~70-line inline LDAP/RMP derivation out of ProfessorPanel into a pure,
unit-testable deriveProfessorViewModel(data) in lib/professorViewModel.ts. The
component keeps only the chrome-dependent photo fallback and settings-driven
section visibility. Behavior preserved.

NEEDS LIVE VERIFICATION (side-panel rendering).
