+++ src/components/professor/ProfessorPanel.tsx
@@ -6,7 +6,7 @@ import { Skeleton } from '@/components/ui/skeleton';
 import { cn } from '@/lib/utils';
 import { useSettings } from '@/lib/hooks/useSettings';
 import { stagger, professorSwitch } from '@/lib/animations';
-import { getFirst } from '@/lib/format';
+import { deriveProfessorViewModel } from '@/lib/professorViewModel';
 import ProfessorHeader from '@/components/professor/ProfessorHeader';
 import ContactInfo from '@/components/professor/ContactInfo';
 import ExpandedDetails from '@/components/professor/ExpandedDetails';
@@ -24,90 +24,48 @@ const GradeDistribution = lazy(() => import('@/components/GradeDistribution'));
  * respects user settings for section visibility, and
  * provides animated professor transitions.
  */
-export default function ProfessorPanel({
-  apiData,
-  rateMyProfessor,
-  reviews,
-  localResearchTopic,
-  localClassesTaught,
-  instructorName,
-  course,
-}: ProfessorData) {
+export default function ProfessorPanel(data: ProfessorData) {
+  const {
+    apiData,
+    rateMyProfessor,
+    reviews,
+    localResearchTopic,
+    instructorName,
+    course,
+  } = data;
   const { settings, loading: settingsLoading } = useSettings();
 
   if (!apiData && !rateMyProfessor) return null;
 
-  // Derive display data from raw API/LDAP fields
-
-  const name = getFirst(apiData?.cn) || instructorName || 'Unknown Professor';
-  const department = getFirst(apiData?.ucscpersonpubdepartmentnumber);
-  const division = getFirst(apiData?.ucscpersonpubdivision);
-  const email = getFirst(apiData?.mail);
-  const phone = getFirst(apiData?.telephonenumber);
-  const officeHours = getFirst(apiData?.ucscpersonpubofficehours);
-  const researchInterest = getFirst(apiData?.ucscpersonpubresearchinterest);
-  const courses = localClassesTaught || apiData?.ucscpersonpubfacultycourses;
-
-  // Photo URL
-  const photoURL = apiData?.jpegphoto;
-  const photoSrc =
-    photoURL && (photoURL as string).includes('uid')
-      ? photoURL
-      : typeof chrome !== 'undefined' && chrome.runtime?.getURL
-        ? chrome.runtime.getURL('images/default_pfp.png')
-        : null;
-
-  // Website
-  let website: string | null = null;
-  const websiteField = apiData?.ucscpersonpubwebsite;
-  if (Array.isArray(websiteField) && websiteField.length > 0) {
-    const raw = websiteField[0];
-    if (typeof raw === 'string' && raw.trim())
-      website = raw.split(' ')[0].trim();
-  } else if (typeof websiteField === 'string' && websiteField.trim()) {
-    website = websiteField.split(' ')[0].trim();
-  }
-
-  // Publications
-  let publicationLinks: string[] = [];
-  const pubField = apiData?.ucscpersonpubselectedpublication;
-  const extractLinks = (html: unknown): string[] => {
-    if (typeof html !== 'string') return [];
-    const links: string[] = [];
-    const hrefPattern = /href="([^"]+)"/g;
-    let m;
-    while ((m = hrefPattern.exec(html)) !== null) links.push(m[1]);
-    return links;
-  };
-  if (Array.isArray(pubField) && pubField.length > 0) {
-    publicationLinks = extractLinks(pubField[0]);
-  } else if (typeof pubField === 'string') {
-    publicationLinks = extractLinks(pubField);
-  }
-
-  // RMP data
+  // Derive display data from raw API/LDAP/RMP fields.
+  const vm = deriveProfessorViewModel(data);
+  const {
+    name,
+    department,
+    division,
+    email,
+    phone,
+    officeHours,
+    researchInterest,
+    courses,
+    website,
+    publicationLinks,
+    overallRating,
+    difficulty,
+    takeAgainPercent,
+    numRatings,
+    topTags,
+    rmpUrl,
+    professorKey,
+  } = vm;
   const rmpNode = rateMyProfessor;
-  const overallRating = rmpNode?.avgRatingRounded ?? null;
-  const difficulty = rmpNode?.avgDifficultyRounded ?? null;
-  // RMP returns -1 for unknown would-take-again; treat any negative as no data.
-  const rawTakeAgain =
-    rmpNode?.wouldTakeAgainPercentRounded ??
-    rmpNode?.wouldTakeAgainPercent ??
-    null;
-  const takeAgainPercent =
-    typeof rawTakeAgain === 'number' && rawTakeAgain >= 0 ? rawTakeAgain : null;
-  const numRatings = rmpNode?.numRatings ?? 0;
-  const ratingTags = Array.isArray(rmpNode?.teacherRatingTags)
-    ? rmpNode!.teacherRatingTags.filter((t) => t?.tagName)
-    : [];
-  const topTags = ratingTags.slice(0, 5);
-  const legacyId = rmpNode?.legacyId;
-  const rmpUrl = legacyId
-    ? `https://www.ratemyprofessors.com/professor/${legacyId}`
-    : null;
 
-  // Unique professor key for transitions
-  const professorKey = rmpNode?.id || `${name}-${department || 'unknown'}`;
+  // Photo: a real directory photo, or the bundled default avatar.
+  const photoSrc =
+    vm.directoryPhoto ??
+    (typeof chrome !== 'undefined' && chrome.runtime?.getURL
+      ? chrome.runtime.getURL('images/default_pfp.png')
+      : null);
 
   // Section visibility from settings
   const sections: SettingsSections = settingsLoading
